Move Bitcoin amount to separate module
